feat: separate human readable time convertor from game-indicators to human-readable-data class, add total time for game to results screen
